Correctly support trees with only a single leaf
When you had trees with only a single leaf in the model, fastforest
previously gave the wrong results. In this case, the number that encodes
the root node of the tree must indicate that the evaluation should jump
straight into the leaf array. We now use negative numbers for that, just
like we do for the node indices within a tree already. The only
difference is that if the root is in the leaf array, we encode the index
as the index minus one so we don't get an ambiguity for the value zero.

This closes #15.