Fix the performance regressions that were caused by recent commits
There were two commits in the last year that degraded the performance of
binary classification with fastforest:

  1. "more interfaces to softmax" (1503edb)
  2. "Correctly support trees with only a single leaf" (5a784d5)

This first performance regression was caused by giving up the function
optimized for binary classification, assuming that using the more
general function for multiclassification will not have much performance
overhead. That was wrong: binary classification got almost 10 % slower,
so a specialized `FastForest::evaluateBinary` function was brought back.

The second regression was caused by removing the assumption that there
are no single-leaf trees. However, handling the single-leaf case in the
evaluation function caused a significant slowdown by about 30 % in
binary classification. To fix this, single-leaf trees are now approached
in a different manner: their leaf values are summed during construction
of the FastForest, and this sum is added as an offset to the FastForest
response in the evaluation. Then we can once again assume that there are
no single-leaf trees.

There were two complications with this new single-leaf tree approach
for multiclassification:

  1. By ignoring the single-leaf trees when filling the tree structure,
     the tree index information is lost, which is important for
     multiclassification evaluation.
  2. We need to track different offsets for the different pair-wise BDTs
     in the multiclassification.

These issues were addressed by adding new data members to the
FastForest class.