Encode correct base response already when reading the model
Encode the base response already when reading the model, depending on
the xgboost version so that we ensure correct inference results.

To ensure no errors with version mismatch can happen, the user is now
forced to add a version hint to the model dump (this will probably be
easier in the future with the JSON parser, where the XGBoost version is
part of the metadata).